Add files via upload
This project Symptom checker is a web application which works on API integration of APIMedic.
This Symptom checker get the input such as Age, Gender and symptom from the user. The diagnosis for the user input 
should be done with the integrated API ( APIMedic ). The authentication key can be get from the "https://apimedic.com/" where you can login and done by the basic 
test you can get the authentication key from the API. After getting the authentication key you need to replace the auth key in the js (javaScript) code then you can easily get the list of symptoms and diagnosis from the API